requirements.txt
gradio==5.23.2
transformers==4.57.0
torch==2.3.1+cpu 
-f https://download.pytorch.org/whl/torch_stable.html
pandas==2.3.3
scikit-learn==1.7.2
gradio-client==1.8.0
huggingface-hub==0.35.3
python-dotenv==1.1.1
requests==2.32.3
